coupang-auto/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
fpdf2>=2.7.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Multi-tenancy and Authentication
psycopg[binary]>=3.2.0
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
